Optimize broadcast frequency for voted players * Also refactored poll components into a new interface
